move SharedDatasetConfig
