+++ illumos-utils/src/zfs.rs
@@ -16,10 +16,10 @@ use camino_tempfile::Utf8TempDir;
 use itertools::Itertools;
 use omicron_common::api::external::ByteCount;
 use omicron_common::disk::CompressionAlgorithm;
-use omicron_common::disk::SharedDatasetConfig;
 use omicron_uuid_kinds::DatasetUuid;
 use rustix::fd::AsRawFd;
 use sled_agent_types::disk::DiskIdentity;
+use sled_agent_types::disk::SharedDatasetConfig;
 use sled_agent_types::inventory::InventoryDataset;
 use slog_error_chain::SlogInlineError;
 use std::collections::BTreeMap;
