+++ nexus/test-utils/src/resource_helpers.rs
@@ -76,7 +76,6 @@ use omicron_common::api::external::RouterRoute;
 use omicron_common::api::external::UserId;
 use omicron_common::disk::DatasetKind;
 use omicron_common::disk::DatasetName;
-use omicron_common::disk::SharedDatasetConfig;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_sled_agent::sim::SledAgent;
 use omicron_test_utils::dev::poll::CondCheckError;
@@ -93,6 +92,7 @@ use oxnet::Ipv6Net;
 use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::OmicronPhysicalDiskConfig;
+use sled_agent_types::disk::SharedDatasetConfig;
 use sled_agent_types::inventory::ZpoolHealth;
 use slog::debug;
 use std::collections::BTreeMap;
