+++ sled-agent/config-reconciler/src/dataset_serialization_task.rs
@@ -30,12 +30,12 @@ use illumos_utils::zfs::Zfs;
 use key_manager_types::VersionedAes256GcmDiskEncryptionKey;
 use omicron_common::disk::DatasetKind;
 use omicron_common::disk::DatasetName;
-use omicron_common::disk::SharedDatasetConfig;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::DatasetUuid;
 use omicron_uuid_kinds::PhysicalDiskUuid;
 use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::disk::DiskIdentity;
+use sled_agent_types::disk::SharedDatasetConfig;
 use sled_agent_types::inventory::InventoryDataset;
 use sled_agent_types::inventory::OrphanedDataset;
 use sled_storage::config::MountConfig;
