+++ sled-agent/config-reconciler/src/reconciler_task/zones.rs
@@ -1304,7 +1304,6 @@ mod tests {
     use omicron_common::address::SLED_PREFIX_LENGTH;
     use omicron_common::disk::DatasetKind;
     use omicron_common::disk::DatasetName;
-    use omicron_common::disk::SharedDatasetConfig;
     use omicron_common::resolvable_files::ResolvableFileSource;
     use omicron_common::update::OmicronInstallManifest;
     use omicron_common::update::OmicronInstallManifestSource;
@@ -1314,6 +1313,7 @@ mod tests {
     use omicron_uuid_kinds::ZpoolUuid;
     use sled_agent_resolvable_files_examples::deserialize_error;
     use sled_agent_types::disk::DatasetConfig;
+    use sled_agent_types::disk::SharedDatasetConfig;
     use sled_agent_types::inventory::OmicronZoneDataset;
     use sled_agent_types::inventory::OmicronZoneImageSource;
     use sled_agent_types::inventory::ZoneKind;
