+++ sled-agent/types/versions/src/latest.rs
@@ -53,6 +53,7 @@ pub mod disk {
     pub use crate::v1::disk::DiskVariant;
     pub use crate::v1::disk::M2Slot;
     pub use crate::v1::disk::OmicronPhysicalDiskConfig;
+    pub use crate::v1::disk::SharedDatasetConfig;
     pub use crate::v1::disk::Zpool;
 }
