+++ sled-diagnostics/src/logs.rs
@@ -985,12 +985,12 @@ mod illumos_tests {
     use illumos_utils::zfs::ZFS;
     use omicron_common::disk::DatasetKind;
     use omicron_common::disk::DatasetName;
-    use omicron_common::disk::SharedDatasetConfig;
     use omicron_common::zpool_name::ZpoolName;
     use omicron_test_utils::dev::test_setup_log;
     use omicron_uuid_kinds::DatasetUuid;
     use omicron_uuid_kinds::ExternalZpoolUuid;
     use sled_agent_types::disk::DatasetConfig;
+    use sled_agent_types::disk::SharedDatasetConfig;
     use tokio::io::AsyncReadExt;
     use tokio::io::AsyncWriteExt;
     use zfs_test_harness::ZfsTestHarness;
